+++ tools/typescript_fixes.js
@@ -1,11 +1,10 @@
-const { readdirSync, readFileSync, writeFileSync, realpathSync, statSync } = require('fs');
+const { readdirSync, readFileSync, writeFileSync, mkdirSync, realpathSync, statSync } = require('fs');
 const path = require('path');
 
 const debug = (namespace, ...args) => {
-    if (
-        process.env.DEBUG_TYPEFIXES
-        && (process.env.DEBUG_TYPEFIXES === '*' || process.env.DEBUG_TYPEFIXES.includes(namespace))
-    ) {
+    const debugTypefixes = process.env.DEBUG_TYPEFIXES;
+
+    if (debugTypefixes && (debugTypefixes === '*' || debugTypefixes.includes(namespace))) {
         console.log(...args);
     }
 };
@@ -15,59 +14,14 @@ const typesPath = './types';
 const paths = {
     main: path.join(typesPath, 'main.d.ts'),
     session_pool: path.join(typesPath, 'session_pool', 'session_pool.d.ts'),
+    utils_log: path.join(typesPath, 'utils_log.d.ts'),
+    playwright_utils: path.join(typesPath, 'playwright_utils.d.ts'),
 };
 
-console.log('Processing', paths['main']);
-
-/**
- * Removes `namespace log` and adds appropriate import and reference to main.d.ts.
- *
- * @param {string[]} input
- * @return {string[]}
- */
-const fixLog = (input) => {
-    const output = [];
-
-    const modes = {
-        base: 0, // Most reading and writing is in `base` mode: keep each line as is.
-        'log-ns-replace': 1, // Erasing log namespace declaration and adding proper import
-    };
-
-    let index = 0;
-    let mode = modes.base;
-    for (const line of input) {
-        try {
-            if (mode === modes.base) {
-                // if (index === 0) {
-                //     output.push('/// <reference path="../src/utils_log.d.ts" />');
-                //     debug('fixLog', 'line #1', mode);
-                // }
-                if (line.match(/^declare namespace log {/)) {
-                    mode = modes['log-ns-replace'];
-                    debug('fixLog', 'switch to', mode, line);
-                } else {
-                    output.push(line);
-                    debug('fixLog', 'out', line);
-                }
-            } else if (mode === modes['log-ns-replace']) {
-                if (line.match(/^\}/)) {
-                    mode = modes.base;
-                    debug('fixLog', 'switch to', mode, line);
-                } else {
-                    // nothing
-                    debug('fixLog', 'remove ', line);
-                }
-            }
-        } finally {
-            index++;
-        }
-    }
-    return output;
-};
+console.log('Processing', paths.main);
 
 const fixSessionPool = async () => {
-    const input = readFileSync(paths.session_pool, { encoding: 'utf8' })
-        .split('\n');
+    const input = readFileSync(paths.session_pool, { encoding: 'utf8' }).split('\n');
 
     const matchLines = [
         /addListener\(event/,
@@ -84,7 +38,7 @@ const fixSessionPool = async () => {
     let changed = false;
 
     for (const line of input) {
-        if (!matchLines.some(s => s.test(line))) {
+        if (!matchLines.some((s) => s.test(line))) {
             output.push(line);
         } else {
             changed = true;
@@ -96,6 +50,90 @@ const fixSessionPool = async () => {
         console.log('Writing file', paths.session_pool);
         writeFileSync(paths.session_pool, output.join('\n'), { encoding: 'utf8' });
     }
+};
+
+const fixUtilsLog = async () => {
+    const input = readFileSync(paths.utils_log, { encoding: 'utf8' }).split('\n');
+    const output = [];
+    let changed = false;
+
+    for (const line of input) {
+        if (!line.includes('import { LoggerOptions } from "@apify/log/log"')) {
+            output.push(line);
+        } else {
+            changed = true;
+            debug('fixUtilsLog', 'removing line', line);
+        }
+    }
+
+    if (changed) {
+        console.log('Writing file', paths.utils_log);
+        writeFileSync(paths.utils_log, output.join('\n'), { encoding: 'utf8' });
+    }
+};
+
+const addTypeReference = (input, types) => {
+    return `${types.map((type) => `/// <reference path="${type}" />`).join('\n')}\n\n${input}`;
+};
+
+const padLeft = (input, length = 4) => {
+    return input.split('\n').map((line) => `${' '.repeat(length)}${line}`).join('\n');
+};
+
+const makePathsAbsolute = (input, root) => {
+    return input.split('\n').map((line) => {
+        return line.replace(/\.\//, `${root}/`);
+    }).join('\n');
+};
+
+const convertPlaywrightIndex = (input) => {
+    const output = makePathsAbsolute(input, 'playwright');
+    return addTypeReference(`declare module 'playwright' {\n${output}\n}`, ['./types/types.d.ts']);
+};
+
+const convertPlaywrightProtocol = (input) => {
+    const output = padLeft(makePathsAbsolute(input, 'playwright/types'));
+    return `declare module 'playwright/types/protocol' {\n${output}\n}`;
+};
+
+const convertPlaywrightStructs = (input) => {
+    const output = padLeft(makePathsAbsolute(input, 'playwright/types'));
+    return addTypeReference(`declare module 'playwright/types/structs' {\n${output}\n}`, ['./types.d.ts']);
+};
+
+const convertPlaywrightTypes = (input) => {
+    const output = padLeft(makePathsAbsolute(input, 'playwright/types'));
+    return addTypeReference(`declare module 'playwright/types/types' {\n${output}\n}`, ['./protocol.d.ts', './structs.d.ts']);
+};
+
+const inlinePlaywrightTypes = async () => {
+    // copy files to `types/playwright`
+    const files = [
+        { path: 'index.d.ts', convertor: convertPlaywrightIndex },
+        { path: 'types/protocol.d.ts', convertor: convertPlaywrightProtocol },
+        { path: 'types/structs.d.ts', convertor: convertPlaywrightStructs },
+        { path: 'types/types.d.ts', convertor: convertPlaywrightTypes },
+    ];
+    mkdirSync(path.join('types', 'playwright'));
+    mkdirSync(path.join('types', 'playwright', 'types'));
+
+    for (const file of files) {
+        const fromPath = path.join('node_modules', 'playwright', file.path);
+        const toPath = path.join('types', 'playwright', file.path);
+        const input = readFileSync(fromPath, { encoding: 'utf8' });
+        const output = file.convertor(input);
+        console.log('Writing file', toPath);
+        writeFileSync(toPath, output, { encoding: 'utf8' });
+    }
+
+    // add type references to playwright_utils.d.ts
+    const input = readFileSync(paths.playwright_utils, { encoding: 'utf8' });
+    const output = addTypeReference(input, [
+        '../types/playwright/index.d.ts',
+        '../types/playwright/types/types.d.ts',
+    ]);
+    console.log('Writing file', paths.playwright_utils);
+    writeFileSync(paths.playwright_utils, output, { encoding: 'utf8' });
 };
 
 /**
@@ -138,7 +176,8 @@ const traverseDirs = async (dir, filter, handleFile) => {
     return types;
 };
 
-const EXCLUDED_FROM_EXPORT = [/^Session$/, /^RequestList$/, /Local$/];
+const EXCLUDED_CLASSES = ['Session', 'RequestList', 'Apify', 'Configuration', 'Configuration', 'BasicCrawler', 'LoggerOptions'];
+const EXCLUDED_FROM_EXPORT = [new RegExp(`^(${EXCLUDED_CLASSES.join('|')})$`), /Local$/];
 
 /**
  * Blindly extracts exported typenames from a `*.d.ts` file.
@@ -150,9 +189,9 @@ const readTypes = (filepath) => {
     const inputByLine = input.split('\n');
     const types = [];
     for (const line of inputByLine) {
-        const matches = line.match(/^export (?:type|class) ([^<\s]+)/); // get exported types without generic Export<T>
+        const matches = line.match(/^export (?:type|class|interface) ([^<\s]+)/); // get exported types without generic Export<T>
         debug('readTypes', !!matches, line);
-        if (matches && !EXCLUDED_FROM_EXPORT.some(s => s.test(matches[1]))) {
+        if (matches && !EXCLUDED_FROM_EXPORT.some((s) => s.test(matches[1]))) {
             types.push(matches[1]);
         }
     }
@@ -206,54 +245,47 @@ const processIndexEsm = async () => {
         },
         readTypes,
     );
-    const input = readFileSync(paths['main'], { encoding: 'utf8' })
+    const input = readFileSync(paths.main, { encoding: 'utf8' })
         .split('\n');
-    console.log('Fixing log module', paths['main']);
-    // const fixedLog = fixLog(input);
-    console.log('Fixing type exports', paths['main']);
+    console.log('Fixing type exports', paths.main);
     const fixedTypes = fixTypes(input, types);
-    console.log('Writing', paths['main']);
-    writeFileSync(paths['main'], fixedTypes.join('\n'));
+    console.log('Writing', paths.main);
+    writeFileSync(paths.main, fixedTypes.join('\n'));
 };
 
 const fixTypesReferences = async () => {
-    await traverseDirs(
-        typesPath,
-        (filename) => {
-            return filename.endsWith('.d.ts');
-        },
-        (filepath) => {
-            const input = readFileSync(filepath, { encoding: 'utf8' })
-                .split('\n');
-            const output = [];
-            let changed = false;
-            let match;
-            for (const line of input) {
-                /* eslint-disable no-cond-assign */
-                if (match = line.match(/\/\/\/\s*<reference\s*types="types-apify\/([^"]+)"\s*\/>/)) {
-                    debug('fixTypesReferences', 'fixing types-apify from file', filepath);
-                    output.push(`/// <reference path="../types-apify/${match[1]}.d.ts" />`);
-                    changed = true;
-                } else if (match = line.match(/^([^"]+)"node\/([^$]+)/)) {
-                    debug('fixTypesReferences', 'fixing "node/" from file', filepath);
-                    output.push(`${match[1]} "${match[2]}`);
-                    changed = true;
-                } else {
-                    output.push(line);
-                }
-                /* eslint-enable no-cond-assign */
+    await traverseDirs(typesPath, (filename) => filename.endsWith('.d.ts'), (filepath) => {
+        const input = readFileSync(filepath, { encoding: 'utf8' }).split('\n');
+        const output = [];
+        let changed = false;
+        let match;
+        for (const line of input) {
+            /* eslint-disable no-cond-assign */
+            if (match = line.match(/\/\/\/\s*<reference\s*types="types-apify\/([^"]+)"\s*\/>/)) {
+                debug('fixTypesReferences', 'fixing types-apify from file', filepath);
+                output.push(`/// <reference path="../types-apify/${match[1]}.d.ts" />`);
+                changed = true;
+            } else if (match = line.match(/^([^"]+)"node\/([^$]+)/)) {
+                debug('fixTypesReferences', 'fixing "node/" from file', filepath);
+                output.push(`${match[1]} "${match[2]}`);
+                changed = true;
+            } else {
+                output.push(line);
             }
+            /* eslint-enable no-cond-assign */
+        }
 
-            if (changed === true) {
-                console.log('Writing', filepath);
-                writeFileSync(filepath, output.join('\n'));
-            }
-        },
-    );
+        if (changed === true) {
+            console.log('Writing', filepath);
+            writeFileSync(filepath, output.join('\n'));
+        }
+    });
 };
 
 (async () => {
     await processIndexEsm();
     await fixTypesReferences();
     await fixSessionPool();
+    await fixUtilsLog();
+    await inlinePlaywrightTypes();
 })();