+++ ts-check/src/index.ts
@@ -1,0 +1,6 @@
+import { Apify } from 'apify';
+
+const sdk = new Apify();
+sdk.main(async () => {
+    // just to see we can actually build such skeleton project...
+}).then();